LayoutTests/platform/gtk/fast/css/getComputedStyle/computed-style-cross-fade-expected.txt

Blocked access to external URL http://example.com/example.png
Blocked access to external URL http://example.com/a.png
Blocked access to external URL http://example.com/b.png
Blocked access to external URL http://example.com/c.png
Blocked access to external URL http://www.iana.org/domains/example/
Blocked access to external URL http://www.iana.org/domains/example/
Blocked access to external URL http://www.iana.org/domains/example/
Blocked access to external URL http://www.iana.org/domains/example/
-webkit-cross-fade


PASS testCrossfade("opacity: 500%", "opacity") is "1"
PASS testCrossfade("background-image: -webkit-cross-fade(url(http://example.com/example.png), url(http://example.com/example.png), 50%)", "background-image") is "-webkit-cross-fade(url(http://example.com/example.png), url(http://example.com/example.png), 0.5)"
PASS testCrossfade("background-image: -webkit-cross-fade(-webkit-cross-fade(url(http://example.com/a.png), url(http://example.com/b.png), 25%), url(http://example.com/example.png), 50%)", "background-image") is "-webkit-cross-fade(-webkit-cross-fade(url(http://example.com/a.png), url(http://example.com/b.png), 0.25), url(http://example.com/example.png), 0.5)"
PASS testCrossfade("background-image: -webkit-cross-fade(url(http://example.com/c.png), -webkit-cross-fade(url(http://example.com/a.png), url(http://example.com/b.png), 25%), 50%)", "background-image") is "-webkit-cross-fade(url(http://example.com/c.png), -webkit-cross-fade(url(http://example.com/a.png), url(http://example.com/b.png), 0.25), 0.5)"
PASS testCrossfade("background-image: -webkit-cross-fade(-webkit-linear-gradient(black, white), url(http://example.com/example.png), 10%)", "background-image") is "-webkit-cross-fade(-webkit-linear-gradient(top, black, white), url(http://example.com/example.png), 0.1)"
PASS testCrossfade("background-image: -webkit-cross-fade(url(http://example.com/example.png), url(http://example.com/example.png), 1.0)", "background-image") is "-webkit-cross-fade(url(http://example.com/example.png), url(http://example.com/example.png), 1)"
PASS testCrossfade("background-image: -webkit-cross-fade(url(http://example.com/example.png), url(http://example.com/example.png), 0.5)", "background-image") is "-webkit-cross-fade(url(http://example.com/example.png), url(http://example.com/example.png), 0.5)"
PASS testCrossfade("background-image: -webkit-cross-fade(url(http://example.com/example.png), url(http://example.com/example.png), 5.0)", "background-image") is "-webkit-cross-fade(url(http://example.com/example.png), url(http://example.com/example.png), 1)"
PASS testCrossfade("background-image: -webkit-cross-fade(url(http://example.com/example.png), url(http://example.com/example.png), 700%)", "background-image") is "-webkit-cross-fade(url(http://example.com/example.png), url(http://example.com/example.png), 1)"
PASS testCrossfade("background-image: -webkit-cross-fade(url(http://example.com/example.png), url(http://example.com/example.png), -20)", "background-image") is "-webkit-cross-fade(url(http://example.com/example.png), url(http://example.com/example.png), 0)"
PASS testCrossfade("background-image: -webkit-cross-fade(url(http://example.com/example.png), url(http://example.com/example.png))", "background-image") is "none"
PASS testCrossfade("background-image: -webkit-cross-fade(url(http://example.com/example.png), url(http://example.com/example.png),)", "background-image") is "none"
PASS testCrossfade("background-image: -webkit-cross-fade(url(http://example.com/example.png))", "background-image") is "none"
PASS testCrossfade("background-image: -webkit-cross-fade(), url(http://example.com/example.png))", "background-image") is "none"
PASS testCrossfade("background-image: -webkit-cross-fade(url(", "background-image") is "none"
PASS testCrossfade("background-image: -webkit-cross-fade(10px 20% 5px #bbb)", "background-image") is "none"
PASS testCrossfade("background-image: -webkit-cross-fade(-webkit-cross-fade(-webkit-cross-fade()))", "background-image") is "none"
PASS testCrossfade("background-image: -webkit-cross-fade(20%, url(http://example.com/a.png), url(http://example.com/b.png))", "background-image") is "none"
PASS testCrossfade("background-image: -webkit-cross-fade(url(http://example.com/a.png),)", "background-image") is "none"
PASS testCrossfade("background-image: -webkit-cross-fade(,)", "background-image") is "none"
PASS testCrossfade("background-image: -webkit-cross-fade(,url(http://example.com/a.png))", "background-image") is "none"
PASS testCrossfade("background-image: -webkit-cross-fade(,,,,,)", "background-image") is "none"
PASS successfullyParsed is true

TEST COMPLETE
